+++ src/main/java/com/example/jomajomadelivery/account/auth/service/UserAuthService.java
@@ -1,10 +1,10 @@
-package com.example.jomajomadelivery.auth.service;
+package com.example.jomajomadelivery.account.auth.service;
 
-import com.example.jomajomadelivery.auth.dto.request.LoginUserDto;
-import com.example.jomajomadelivery.auth.dto.request.SignUpUserDto;
-import com.example.jomajomadelivery.auth.jwt.TokenProvider;
-import com.example.jomajomadelivery.auth.oauth2.SocialProvider;
-import com.example.jomajomadelivery.auth.repository.UserAuthRepository;
+import com.example.jomajomadelivery.account.auth.dto.request.LoginUserDto;
+import com.example.jomajomadelivery.account.auth.dto.request.SignUpUserDto;
+import com.example.jomajomadelivery.account.jwt.TokenProvider;
+import com.example.jomajomadelivery.account.oauth2.service.SocialProvider;
+import com.example.jomajomadelivery.account.auth.repository.UserAuthRepository;
 import com.example.jomajomadelivery.user.domain.Password;
 import com.example.jomajomadelivery.user.entity.User;
 import jakarta.servlet.http.Cookie;
